pyproject.toml
[project]
name = "leetcode-daily-discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apscheduler>=3.10.4",
    "beautifulsoup4>=4.13.3",
    "colorlog>=6.9.0",
    "discord-py>=2.5.2",
    "aiohttp>=3.9.3",
    "python-dotenv>=1.0.0",
    "pytz>=2025.2",
    "requests>=2.32.3",
    "tomli>=2.0.1; python_version < '3.11'",
]
